+++ bundles/org.palladiosimulator.analyzer.accuracy/src/org/palladiosimulator/analyzer/accuracy/transformation/AbstractAccuracyInfluenceSEFFTransformationStrategy.java
@@ -22,7 +22,7 @@ import org.palladiosimulator.analyzer.accuracy.issues.AccuracyIssueFactory;
 import org.palladiosimulator.analyzer.accuracy.jobs.TransformPCMForAccuracyInfluenceAnalysisJob;
 import org.palladiosimulator.analyzer.quality.util.EMFHelper;
 import org.palladiosimulator.analyzer.quality.util.QualityAnnotationLookup;
-import org.palladiosimulator.analyzer.workflow.blackboard.PCMResourceSetPartition;
+import org.palladiosimulator.analyzer.workflow.core.blackboard.PCMResourceSetPartition;
 import org.palladiosimulator.analyzer.workflow.core.runconfig.AccuracyInfluenceAnalysisState;
 import org.palladiosimulator.pcm.seff.AbstractAction;
 import org.palladiosimulator.pcm.seff.AbstractBranchTransition;
