+++ bundles/org.palladiosimulator.analyzer.workflow.core/META-INF/MANIFEST.MF
@@ -4,6 +4,7 @@ Bundle-Name: PCM Workflow Jobs and Configurations Core
 Bundle-SymbolicName: org.palladiosimulator.analyzer.workflow.core
 Bundle-Version: 5.2.1.qualifier
 Export-Package: org.palladiosimulator.analyzer.workflow.core,
+ org.palladiosimulator.analyzer.workflow.core.blackboard,
  org.palladiosimulator.analyzer.workflow.core.configurations,
  org.palladiosimulator.analyzer.workflow.core.jobs,
  org.palladiosimulator.analyzer.workflow.core.runconfig
@@ -17,7 +18,8 @@ Require-Bundle: org.apache.log4j;bundle-version="1.2.19",
  org.palladiosimulator.analyzer.completions;bundle-version="5.2.1",
  de.uka.ipd.sdq.workflow.mdsd.core;bundle-version="5.2.1",
  de.uka.ipd.sdq.workflow.logging.console;bundle-version="5.2.1",
- de.uka.ipd.sdq.workflow.mdsd.blackboard;bundle-version="5.2.1"
+ de.uka.ipd.sdq.workflow.mdsd.blackboard;bundle-version="5.2.1",
+ de.uka.ipd.sdq.featureconfig;bundle-version="5.2.1"
 Bundle-Vendor: palladiosimulator.org
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
